[FEAT] 주문 C 구현

+++ src/main/java/hoxy/hLivv/dto/MemberCouponDto.java
@@ -17,7 +17,7 @@ public class MemberCouponDto {
     @Temporal(TemporalType.DATE)
     private LocalDate expireDate;
     private String couponDesc;
-    private Double discountRate;
+    private Integer discountRate;
     private boolean isUsed;
 
     public static MemberCouponDto from(MemberCoupon memberCoupon) {
